Denoiser separation
