+++ app/processors/models_data.py
@@ -1,5 +1,6 @@
 import os
 from pathlib import Path
+import numpy as np
 
 models_dir = Path(__file__).resolve().parent.parent.parent / "model_assets"
 # ensure ref-ldm paths exist
@@ -17,6 +18,177 @@ for _subfolder in ("liveportrait_onnx", "performrecast_onnx"):
     os.makedirs(models_dir / _subfolder, exist_ok=True)
 
 assets_repo = "https://github.com/visomaster/visomaster-assets/releases/download"
+
+ARCFACE_DST = np.array(
+    [
+        [38.2946, 51.6963],
+        [73.5318, 51.5014],
+        [56.0252, 71.7366],
+        [41.5493, 92.3655],
+        [70.7299, 92.2041],
+    ],
+    dtype=np.float32,
+)
+
+FFHQ_KPS = np.array(
+    [
+        [192.98138, 239.94708],
+        [318.90277, 240.1936],
+        [256.63416, 314.01935],
+        [201.26117, 371.41043],
+        [313.08905, 371.15118],
+    ],
+    dtype=np.float32,
+)
+
+LANDMARKS_SUBSET_IDXS = [
+    0,
+    1,
+    4,
+    5,
+    6,
+    7,
+    8,
+    10,
+    13,
+    14,
+    17,
+    21,
+    33,
+    37,
+    39,
+    40,
+    46,
+    52,
+    53,
+    54,
+    55,
+    58,
+    61,
+    63,
+    65,
+    66,
+    67,
+    70,
+    78,
+    80,
+    81,
+    82,
+    84,
+    87,
+    88,
+    91,
+    93,
+    95,
+    103,
+    105,
+    107,
+    109,
+    127,
+    132,
+    133,
+    136,
+    144,
+    145,
+    146,
+    148,
+    149,
+    150,
+    152,
+    153,
+    154,
+    155,
+    157,
+    158,
+    159,
+    160,
+    161,
+    162,
+    163,
+    168,
+    172,
+    173,
+    176,
+    178,
+    181,
+    185,
+    191,
+    195,
+    197,
+    234,
+    246,
+    249,
+    251,
+    263,
+    267,
+    269,
+    270,
+    276,
+    282,
+    283,
+    284,
+    285,
+    288,
+    291,
+    293,
+    295,
+    296,
+    297,
+    300,
+    308,
+    310,
+    311,
+    312,
+    314,
+    317,
+    318,
+    321,
+    323,
+    324,
+    332,
+    334,
+    336,
+    338,
+    356,
+    361,
+    362,
+    365,
+    373,
+    374,
+    375,
+    377,
+    378,
+    379,
+    380,
+    381,
+    382,
+    384,
+    385,
+    386,
+    387,
+    388,
+    389,
+    390,
+    397,
+    398,
+    400,
+    402,
+    405,
+    409,
+    415,
+    454,
+    466,
+    468,
+    469,
+    470,
+    471,
+    472,
+    473,
+    474,
+    475,
+    476,
+    477,
+]
 
 arcface_mapping_model_dict = {
     "Inswapper128": "Inswapper128ArcFace",
